+++ lib/src/types/properties.ts
@@ -517,7 +517,7 @@ export interface MapProperty<T extends Record<string, CMSType> = Record<string, 
 
     /**
      * Render this map as a key-value table that allows to use
-     * arbitrary keys.
+     * arbitrary keys. You don't need to define the properties in this case.
      */
     keyValue?: boolean;